pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grafinder"
version = "0.1.0"
description = "Local keyword-driven data discovery agent with Grafana visualization."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "asyncpg>=0.29.0",
  "beautifulsoup4>=4.12.3",
  "crawl4ai>=0.4.0",
  "duckduckgo-search>=6.2.13",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "openai>=1.40.0",
  "playwright>=1.40.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-dateutil>=2.9.0.post0",
  "sqlalchemy>=2.0.32",
  "uvicorn[standard]>=0.30.6",
]

[project.scripts]
grafinder = "app.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["templates/*.html"]
